+++ dsl/src/test/groovy/com/readytalk/jenkins/model/types/WorkflowTest.groovy
@@ -6,8 +6,10 @@ import com.readytalk.jenkins.model.YamlParser
 class WorkflowTest extends ModelSpecification {
   def setup() {
     types {
-      add WorkflowSettings.instance
-      job('testJob', ['base', 'workflowSettings'])
+      add MultibranchSettings.instance
+      add PipelineSettings.instance
+      job('multibranchJob', ['base', 'multibranchSettings'])
+      job('pipelineJob', ['base', 'pipelineSettings'])
     }
   }
 
@@ -18,7 +20,7 @@ class WorkflowTest extends ModelSpecification {
     name: workflow-thing
     base:
       type: multibranchWorkflowJob
-    workflowSettings:
+    multibranchSettings:
       remote: 'fake-repo'
 """))
 
@@ -26,5 +28,48 @@ class WorkflowTest extends ModelSpecification {
 
     then:
     job.name() == 'org.jenkinsci.plugins.workflow.multibranch.WorkflowMultiBranchProject'
+  }
+
+  def "Can set credentials via dsl block"() {
+    when:
+    def jobs = generate(YamlParser.parse("""
+- multibranchJob:
+    name: workflow-thing
+    base:
+      type: multibranchWorkflowJob
+    multibranchSettings:
+      remote: 'fake-repo'
+      dsl: |
+        credentialsId('fake-creds-id')
+"""))
+
+    Node job = jobs.find { it.name == 'workflow-thing' }.getNode()
+
+    then:
+    //Should only be one BranchSource with this setup
+    job.sources.data.size() == 1
+    job.sources.data.'jenkins.branch.BranchSource'[0].source.credentialsId[0].value() == 'fake-creds-id'
+  }
+
+  def "Can set inline pipeline script"() {
+    when:
+    String script = "echo 'Hello world!'"
+    def jobs = generate(YamlParser.parse("""
+- pipelineJob:
+    name: pipeinline
+    base:
+      type: workflowJob
+    pipelineSettings:
+      script: |
+        ${script}
+      groovySandbox: true
+"""))
+
+    Node pipeinline = jobs.find { it.name == 'pipeinline' }.getNode()
+
+
+    then:
+    pipeinline.definition.script[0].value().trim() == script
+    pipeinline.definition.sandbox[0].value() == true
   }
 }